+++ piecrust/src/session.rs
@@ -12,7 +12,9 @@ use std::sync::{mpsc, Arc};
 
 use bytecheck::CheckBytes;
 use dusk_wasmtime::{Engine, LinearMemory, MemoryCreator, MemoryType};
-use piecrust_uplink::{ContractId, Event, ARGBUF_LEN, SCRATCH_BUF_BYTES};
+use piecrust_uplink::{
+    ContractId, EconomicMode, Event, ARGBUF_LEN, SCRATCH_BUF_BYTES,
+};
 use rkyv::ser::serializers::{
     BufferScratch, BufferSerializer, CompositeSerializer,
 };
@@ -381,7 +383,7 @@ impl Session {
             return Err(InitalizationError("init call not allowed".into()));
         }
 
-        let (data, gas_spent, call_tree) =
+        let (data, gas_spent, call_tree, economic_mode) =
             self.call_inner(contract, fn_name, fn_arg.into(), gas_limit)?;
         let events = mem::take(&mut self.inner.events);
 
@@ -391,6 +393,7 @@ impl Session {
             events,
             call_tree,
             data,
+            economic_mode,
         })
     }
 
@@ -748,7 +751,7 @@ impl Session {
         fname: &str,
         fdata: Vec<u8>,
         limit: u64,
-    ) -> Result<(Vec<u8>, u64, CallTree), Error> {
+    ) -> Result<(Vec<u8>, u64, CallTree, EconomicMode), Error> {
         let stack_element = self.push_callstack(contract, limit)?;
         let instance = self
             .instance(&stack_element.contract_id)
@@ -762,6 +765,7 @@ impl Session {
             })?;
 
         let arg_len = instance.write_bytes_to_arg_buffer(&fdata)?;
+        instance.clear_eco_mode();
         let ret_len = instance
             .call(fname, arg_len, limit)
             .map_err(|err| {
@@ -777,6 +781,7 @@ impl Session {
             })
             .map_err(Error::normalize)?;
         let ret = instance.read_bytes_from_arg_buffer(ret_len as u32);
+        let economic_mode = instance.read_eco_mode();
 
         let spent = limit - instance.get_remaining_gas();
 
@@ -797,7 +802,7 @@ impl Session {
         mem::swap(&mut self.inner.call_tree, &mut call_tree);
         call_tree.update_spent(spent);
 
-        Ok((ret, spent, call_tree))
+        Ok((ret, spent, call_tree, economic_mode))
     }
 
     pub fn contract_metadata(
@@ -827,6 +832,9 @@ pub struct CallReceipt<T> {
 
     /// The data returned by the called contract.
     pub data: T,
+
+    /// Economic mode applied during the execution of the call.
+    pub economic_mode: EconomicMode,
 }
 
 impl CallReceipt<Vec<u8>> {
@@ -847,6 +855,7 @@ impl CallReceipt<Vec<u8>> {
             events: self.events,
             call_tree: self.call_tree,
             data,
+            economic_mode: self.economic_mode,
         })
     }
 }